+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Breaking Changes
 
 ### Features
+- Onboard z_score normalization and RRF combination to HYBRID_OPTIMIZER experiment ([#465](https://github.com/opensearch-project/search-relevance/pull/465))
 
 ### Enhancements